# docs/contracts/specs/rest.openapi.yaml
openapi: 3.1.0
info:
  title: MyRoboTaxi Telemetry REST API
  version: 1.0.0
  summary: REST surface for the MyRoboTaxi SDK v1.
  description: |
    Non-streaming REST surface that binds the Go telemetry server's REST
    handlers (PLANNED, see divergences DV-19/DV-20) to the MyRoboTaxi
    TypeScript and Swift SDKs. This document is the machine-readable twin
    of `docs/contracts/rest-api.md`. The markdown is the human source of
    truth; this YAML is consumed by OpenAPI tooling, `contract-tester`
    fixtures, and SDK code generators.

    The snapshot response references `schemas/vehicle-state.schema.json`
    via `$ref` rather than re-declaring the VehicleState shape. REST-only
    shapes (DriveSummary, DriveDetail, DriveRoute, RoutePoint, Invite,
    CreateInviteRequest, ErrorEnvelope, PaginatedDrives, PaginatedInvites,
    DeleteUserResponse) are declared inline under components/schemas for
    the v1 scope. A follow-up will extract them to sibling JSON Schemas.

    Anchored requirements: FR-3.2, FR-3.3, FR-3.4, FR-5.1, FR-5.2, FR-5.3,
    FR-5.4, FR-5.5, FR-6.1, FR-6.2, FR-7.1, FR-7.3, FR-9.1, FR-9.2, FR-10.1,
    FR-10.2, NFR-3.5, NFR-3.11, NFR-3.19, NFR-3.20, NFR-3.21, NFR-3.22,
    NFR-3.23, NFR-3.27, NFR-3.29.

    Known, accepted divergences between this spec and the current
    `internal/server/` + `internal/store/` implementation are catalogued
    in rest-api.md §10 with stable DV-NN identifiers (DV-19 through DV-23
...
  contact:
    name: sdk-architect
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /vehicles/{vehicleId}/snapshot:
    get:
      operationId: getVehicleSnapshot
      tags: [snapshot]
      summary: Cold-load vehicle state snapshot.
      description: |
        Returns the full current VehicleState for a single vehicle. This
        is the snapshot the SDK fetches on cold page load and on every
        reconnect per NFR-3.11 (see websocket-protocol.md §7.2 reconnect
        sequence). The snapshot is the DB source-of-truth
        (data-lifecycle.md §1.2); the WebSocket is the real-time channel.

        Latency target: p95 < 500 ms end-to-end per requirements.md §3.1.

        RBAC: owners see all fields in vehicle-state.schema.json; viewers
        see all fields EXCEPT licensePlate (forward-looking mask -- see
        rest-api.md §5.2.1). The limited_viewer extension slot (FR-5.5)
        is NOT implemented in v1 but is documented as the extension seam.
      x-anchored-requirements:
        - NFR-3.5
        - NFR-3.11
        - FR-1.1
        - FR-1.2
        - FR-2.1
      x-divergence: DV-20
      parameters:
        - $ref: '#/components/parameters/VehicleIdPath'
      responses:
        '200':
          description: Current VehicleState snapshot.
          content:
            application/json:
              schema:
                $ref: 'https://myrobotaxi.com/schemas/vehicle-state.schema.json'
              examples:
                parked-vehicle:
                  summary: Parked vehicle, no active navigation, disconnected from charger.
                  value:
                    vehicleId: clxyz1234567890abcdef
                    name: Stumpy
                    model: Model 3
                    year: 2024
                    color: Midnight Silver Metallic
                    status: parked
                    speed: 0
                    heading: 180
                    latitude: 10.0
                    longitude: 20.0
                    locationName: Home
                    locationAddress: 123 Market St, San Francisco, CA
                    gearPosition: P
                    chargeLevel: 78
                    chargeState: Disconnected
                    estimatedRange: 245
                    timeToFull: null
                    interiorTemp: 68
                    exteriorTemp: 55
                    odometerMiles: 12458
                    fsdMilesSinceReset: 412.7
                    destinationName: null
                    destinationAddress: null
                    destinationLatitude: null
                    destinationLongitude: null
                    originLatitude: null
                    originLongitude: null
                    etaMinutes: null
                    tripDistanceRemaining: null
                    navRouteCoordinates: null
                    lastUpdated: '2026-04-13T18:22:01Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/VehicleNotOwned'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      x-role-masks:
        owner: ['*']
        viewer:
          description: All VehicleState fields except licensePlate (forward-looking; licensePlate is not currently a member of VehicleState).
          excludes:
            - licensePlate
        limited_viewer:
          description: FR-5.5 extension seam -- NOT implemented in v1. Documented for future readiness.
          excludes:
            - licensePlate
            - navRouteCoordinates
            - destinationName
            - destinationAddress
            - destinationLatitude
            - destinationLongitude
            - originLatitude
            - originLongitude
          modifications:
            latitude: coarse_hash
            longitude: coarse_hash
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer-style session token resolved by the consumer's
        getToken() callback (FR-6.1). The token is the SAME value the
        SDK passes in the WebSocket auth frame (websocket-protocol.md
        §2.2) -- the SDK maintains a single credential surface across
        WS and REST.

        Unlike the WebSocket path (which puts the token in the first
        WebSocket frame because browsers cannot set arbitrary headers
        on an upgrade), REST uses the standard Authorization header
        because REST has no upgrade handshake.

        On 401 the SDK MUST call getToken() to refresh and retry the
        request exactly once per FR-6.2. A second 401 is terminal and
        the error surfaces to the consumer as a typed auth_failed
        error.

        The token is P1 per data-classification.md §1.2. The server
        MUST NOT log the Authorization header in any structured log
        field, error message, metric label, or crash report.
  parameters:
    VehicleIdPath:
      name: vehicleId
      in: path
      required: true
      description: Opaque DB cuid for the vehicle (FR-4.2). Never a VIN.
      schema:
        type: string
        pattern: '^c[a-z0-9]+$'
        minLength: 20
        maxLength: 40
      example: clxyz1234567890abcdef
...
  responses:
    Unauthorized:
      description: |
        Missing, malformed, or invalid bearer token. On receipt the SDK
        MUST call getToken() and retry once per FR-6.2. A second 401 is
        terminal and surfaces as a typed auth_failed error.

        Special case: 401 auth_failed with subCode=reauth_required is
        emitted by /users/me (DELETE) and /users/me/export (GET) when
        the bearer token is valid but the caller's auth_time claim is
        outside the recent-auth window (rest-api.md §7.6 / §7.7,
        MYR-76). This subCode is carved out from the getToken() retry
        path — the SDK MUST surface it to the consumer's auth layer
        for an interactive sign-in flow rather than swallowing it with
        a silent refresh, because the auth_time claim only advances on
        a fresh OAuth round-trip.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            auth-failed:
              value:
                error:
                  code: auth_failed
                  message: invalid token
                  subCode: null
            auth-timeout:
              value:
                error:
                  code: auth_timeout
                  message: 'authenticator.ValidateToken: deadline exceeded'
                  subCode: null
            reauth-required:
              summary: Recent-login re-auth gate failed (rest-api.md §7.6 / §7.7, MYR-76)
              value:
                error:
                  code: auth_failed
                  message: recent re-authentication required
                  subCode: reauth_required
    VehicleNotOwned:
      description: Caller is not the owner or an accepted viewer of the requested vehicleId (or drive's parent vehicleId).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            vehicle-not-owned:
              value:
                error:
                  code: vehicle_not_owned
                  message: caller does not own this vehicle
...
    NotFound:
      description: |
        The requested resource does not exist, or it exists but is not
        visible to the caller. REST-only error code -- the WebSocket
        path enforces ownership via silent filtering in Hub.Broadcast
        (websocket-protocol.md §4.5) and has no request-response
        equivalent. Added to the shared ErrorPayload.code enum as a
        REST-only extension -- see rest-api.md §4.1.1.a and DV-20.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            unknown-vehicle:
              value:
                error:
                  code: not_found
                  message: vehicle does not exist
            unknown-drive:
              value:
                error:
                  code: not_found
                  message: drive does not exist
...
    RateLimited:
      description: |
        Per-user REST rate limit breached (PLANNED: 120 requests/minute,
        see DV-22). SDK MUST apply exponential backoff per
        websocket-protocol.md §7.1 reconnect curve.
      headers:
        Retry-After:
          description: Advisory delay in seconds before the SDK should retry.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            rest-rate-limited:
              value:
                error:
                  code: rate_limited
                  message: REST rate limit exceeded
...
    InternalError:
      description: Unexpected server failure. SDK auto-retries with exponential backoff up to 3 attempts before surfacing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            catchall:
              value:
                error:
                  code: internal_error
                  message: internal server error
  schemas:
    ErrorEnvelope:
      type: object
      description: |
        Standard error envelope for all non-2xx REST responses. The
        `error.code` field is a stable typed enum that SDK consumers
        branch on per FR-7.1. The `error.message` field is a
        human-readable description for logs and developer tooling and
        MUST NOT contain any P1 value (no GPS, no addresses, no tokens,
        no emails, no raw VINs). See rest-api.md §4.1 and
        data-classification.md §2.2.

        The code enum is a superset of the WebSocket ErrorPayload.code
        enum in schemas/ws-messages.schema.json. REST-only codes
        (not_found, invalid_request, service_unavailable) are called
        out in the value descriptions. Consumers that share a CoreError
        type across transports reuse the same enum values as the WS
        side -- the enums intentionally overlap.
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          description: Nested error object carrying the typed code, human-readable message, and optional sub-code. All inner fields are P0; this wrapper is P0 overall.
          x-classification: P0
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: |
                Stable typed error code. Reuses the WebSocket
                ErrorPayload.code catalog plus the REST-only extensions
                not_found, invalid_request, and service_unavailable.
                SDK consumers branch on this value, never on `message`.
              enum:
                - auth_failed
                - auth_timeout
                - permission_denied
                - vehicle_not_owned
                - not_found
                - invalid_request
                - rate_limited
                - internal_error
                - service_unavailable
              x-classification: P0
            message:
              type: string
              description: |
                Human-readable description for logs and developer
                tooling. P0 but MUST NOT contain P1 values -- no GPS,
                no addresses, no tokens, no emails, no VINs (unless
                redacted to ***XXXX via redactVIN()). Error
                construction sites use opaque IDs (vehicleId, driveId,
                userId, inviteId) for correlation.
              x-classification: P0
            subCode:
              type: string
              nullable: true
              description: |
                Optional typed sub-code. The field is always present on
                the wire; the carrier emits JSON `null` when there is
                no sub-code to attach. Two non-null values in v1:

                - device_cap (WS-only): per-user concurrent-connection
                  cap breach on the WebSocket path. Declared on the
                  REST envelope only for shared-type compatibility
                  with the SDK's CoreError union — REST v1 does not
                  emit this value.
                - reauth_required (REST-only): the recent-login
                  re-auth gate on DELETE /users/me and GET
                  /users/me/export rejected the request because the
                  caller's auth_time claim is missing or older than
                  REAUTH_MAX_AGE_SEC (default 300 s). The SDK MUST
                  surface this subCode to the consumer's auth layer
                  for an interactive sign-in flow; see rest-api.md
                  §7.6 / §7.7 (MYR-76) and the §4.1.1 auth_failed
                  row.
              enum: [device_cap, reauth_required]
              x-classification: P0
...